+++ lmcache/v1/compute/models/utils.py
@@ -25,6 +25,11 @@ def infer_model_from_vllm(vllm_model, blender, enable_sparse: bool = False):
 
         return LMCQwenModel(vllm_model, blender, enable_sparse)
 
+    elif model_name == "Qwen3ForCausalLM":
+        # First Party
+        from lmcache.v1.compute.models.qwen3 import LMCQwen3Model
+
+        return LMCQwen3Model(vllm_model, blender, enable_sparse)
     else:
         # TODO(Jiayi): Add support for more models
         raise NotImplementedError(
